% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHSIOptions.R
\name{getHSIOptions}
\alias{getHSIOptions}
\title{An API to Download HSI Options Data}
\usage{
getHSIOptions()
}
\description{
\code{getHSIOptions} returns HSI options in dataframe.
}
\details{
The function is to retrieve HSI options data from the HKEX website.
}
\examples{
\dontrun{
   getHSIOptions()
}
}
\references{
https://www.hkex.com.hk
}
\keyword{hkex.api}

