% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastShortSelling.R
\name{getLastShortSelling}
\alias{getLastShortSelling}
\title{An API to Download Short Selling Turnover on Main Board}
\usage{
getLastShortSelling()
}
\description{
\code{getLastShortSelling} returns a data frame.
}
\details{
The function is to retrieve short selling turnover on main board
}
\examples{
\dontrun{
   getLastShortSelling()
}
}
\references{
https://www.hkex.com.hk
}
\keyword{hkex.api}

