% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepInsertSQL.R
\name{prepInsertSQL}
\alias{prepInsertSQL}
\title{A Function to Generate a List of SQL statements from Dataframe.}
\usage{
prepInsertSQL(tableName = "", df = NULL, quote = NULL)
}
\arguments{
\item{tableName}{a table name}

\item{df}{dataframe that provides data}

\item{quote}{(Optional) it is a vector indicates whether a data element should be quoted or not. 1 means quoted while 0 means not.}
}
\description{
\code{prepInsertSQL} returns a list of insert SQL statements
}
\details{
The function is to generate insert SQL statements with data.frame data
}
\examples{
\dontrun{
   testData=data.frame(name=c("peter","john"), age=c(18,23))
   sqlStatement=prepInsertSQL(tableName="student", testData, quote=c(1,0))
}

}
\keyword{hkex.api}

