% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHSIOptions.R
\name{getHSIOptions}
\alias{getHSIOptions}
\title{An API to Download HSI Options Data}
\usage{
getHSIOptions()
}
\arguments{
\item{no}{parameter}
}
\description{
The function is to retrieve HSI options data from the HKEX website.
}
\examples{
data <- getHSIOptions()
getHSIOptions()
}
\references{
https://www.hkex.com.hk
}
\keyword{hkex.api}

